Added PB range Off, 1-12 semitones
The pitchbend range was 2 semitones and fixed, I changed it to Off, 1, 2. But I noticed that the tunings were slightly off, so I created a new maps for the PB ranges, doubling from 1.00 - 2.00 on the positive bend and 1.00 - 0.50 on the negative bend range. This fixed the accuracy of the bend ranges and allow me to implement 0-12 semitones.